+++ custom_components/mydolphin_plus/component/helpers/__init__.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 import logging
 import sys
 
@@ -49,3 +50,30 @@ def clear_ha(hass: HomeAssistant, entry_id):
         hass.data[DATA] = dict()
 
     del hass.data[DATA][entry_id]
+
+
+def get_cleaning_mode_details(mode):
+    details = CLEANING_MODES.get(mode)
+
+    return details
+
+
+def get_cleaning_mode_name(mode):
+    details = get_cleaning_mode_details(mode)
+    details_parts = details.split(" - ")
+    name = details_parts[0]
+
+    return name
+
+
+def get_cleaning_mode_description(mode):
+    details = get_cleaning_mode_details(mode)
+    details_parts = details.split(" - ")
+    description = details_parts[1]
+
+    return description
+
+
+def get_date_time_from_timestamp(timestamp):
+    result = datetime.fromtimestamp(timestamp)
+    return result