+++ custom_components/mydolphin_plus/component/managers/home_assistant.py
@@ -6,16 +6,17 @@ https://home-assistant.io/components/mydolphin_plus/
 from __future__ import annotations
 
 import calendar
+import datetime
 import logging
 import sys
 
+from homeassistant.components.binary_sensor import BinarySensorDeviceClass
 from homeassistant.components.sensor import SensorDeviceClass, SensorStateClass
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 
 from ...component.api.mydolphin_plus_api import MyDolphinPlusAPI
 from ...component.helpers.const import *
-from ...component.helpers.enums import ConnectivityStatus
 from ...component.managers.event_manager import MyDolphinPlusEventManager
 from ...configuration.managers.configuration_manager import (
     ConfigurationManager,
@@ -23,6 +24,11 @@ from ...configuration.managers.configuration_manager import (
 )
 from ...configuration.models.config_data import ConfigData
 from ...core.managers.home_assistant import HomeAssistantManager
+from ..helpers import (
+    get_cleaning_mode_description,
+    get_cleaning_mode_name,
+    get_date_time_from_timestamp,
+)
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -89,16 +95,70 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
 
             _LOGGER.error(f"Failed to async_update_data_providers, Error: {ex}, Line: {line_number}")
 
+    def register_services(self, entry: ConfigEntry | None = None):
+        services = {
+            SERVICE_NAVIGATE: {
+                "handler": self.async_service_navigate,
+                "scheme": SERVICE_SCHEMA_NAVIGATE
+            },
+            SERVICE_PICKUP: {
+                "handler": self.async_service_pickup,
+                "scheme": SERVICE_SCHEMA_NAVIGATE
+            },
+            SERVICE_SET_DAILY_SCHEDULE: {
+                "handler": self.async_service_set_daily_schedule,
+                "scheme": SERVICE_SCHEMA_DAILY_SCHEDULE
+            },
+            SERVICE_DELAYED_CLEAN: {
+                "handler": self.async_service_delayed_clean,
+                "scheme": SERVICE_SCHEMA_DELAYED_CLEAN
+            }
+        }
+
+        for service in services:
+            service_details = services.get(service)
+            handler = service_details.get("handler")
+            scheme = service_details.get("scheme")
+
+            self._hass.services.async_register(DOMAIN, service, handler, scheme)
+
+    async def async_service_navigate(self, call_service):
+        device = call_service.get("device")
+        direction = call_service.get("direction")
+
+        await self.drive(device, direction)
+
+    async def async_service_pickup(self, call_service):
+        device = call_service.get("device")
+
+        await self.pickup(device)
+
+    async def async_service_set_daily_schedule(self, call_service):
+        device = call_service.get("device")
+        day = call_service.get("day")
+        enabled = call_service.get("enabled")
+        cleaning_mode = call_service.get("mode", "all")
+        job_time = call_service.get("time")
+
+        await self.set_schedule(device, day, enabled, cleaning_mode, job_time)
+
+    async def async_service_delayed_clean(self, call_service):
+        device = call_service.get("device")
+        enabled = call_service.get("enabled")
+        cleaning_mode = call_service.get("mode", "all")
+        job_time = call_service.get("time")
+
+        await self.set_delay(device, enabled, cleaning_mode, job_time)
+
     def load_entities(self):
         data = self._api.data
         name = data.get("Robot Name")
 
         self._load_select_cleaning_mode(name, data)
         self._load_select_led_mode(name, data)
-        self._load_remote(name, data)
+        self._load_binary_sensor_connection(name, data)
         self._load_binary_sensor_status(name, data)
         self._load_binary_sensor_filter_status(name, data)
-        self._load_sensor_connection_type(name, data)
         self._load_sensor_cleaning_time(name, data)
         self._load_sensor_cleaning_time_left(name, data)
         self._load_switch_power(name, data)
@@ -122,14 +182,25 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Cleaning Mode"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            cycle_info = data.get("cycleInfo", {})
+            cleaning_mode = cycle_info.cleaningMode("cleaningMode", {})
+            mode = cleaning_mode.cleaningMode("mode", "all")
+            mode_name = get_cleaning_mode_name(mode)
+            mode_description = get_cleaning_mode_description(mode)
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            cycle_time_minutes = cleaning_mode.cleaningMode("cycleTime", 0)
+            cycle_start_time_ts = cycle_info.cleaningMode("cycleStartTime", 0)
+            cycle_start_time = get_date_time_from_timestamp(cycle_start_time_ts)
+
+            cycle_time = str(datetime.timedelta(minutes=cycle_time_minutes))
+
+            state = mode_name
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Description": mode_description,
+                "Cycle time": cycle_time,
+                "Start time": cycle_start_time
+            }
 
             entity = self.entity_manager.get(DOMAIN_SELECT, entity_name)
             created = entity is None
@@ -144,7 +215,6 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
 
             data = {
                 "state": (entity.state, str(state)),
-                "attributes": (entity.attributes, attributes),
                 "device_name": (entity.device_name, device),
             }
 
@@ -170,14 +240,17 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Led Mode"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            led = data.get("led", {})
+            led_mode = led.cleaningMode("ledMode", 1)
+            led_intensity = led.cleaningMode("ledIntensity", 80)
+            led_enable = led.cleaningMode("ledEnable", False)
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            state = led_mode
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Enable": led_enable,
+                "ledIntensity": led_intensity
+            }
 
             entity = self.entity_manager.get(DOMAIN_SELECT, entity_name)
             created = entity is None
@@ -192,7 +265,6 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
 
             data = {
                 "state": (entity.state, str(state)),
-                "attributes": (entity.attributes, attributes),
                 "device_name": (entity.device_name, device),
             }
 
@@ -210,54 +282,6 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
                 ex, f"Failed to load {DOMAIN_SELECT}: {entity_name}"
             )
 
-    def _load_remote(
-            self,
-            device: str,
-            data: dict
-    ):
-        entity_name = f"{device} Remote"
-
-        try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
-
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
-
-            entity = self.entity_manager.get(DOMAIN_REMOTE, entity_name)
-            created = entity is None
-
-            if created:
-                entity = self.entity_manager.get_empty_entity(self.entry_id)
-
-                entity.id = entity_name
-                entity.name = entity_name
-                entity.icon = DEFAULT_ICON
-                entity.domain = DOMAIN_REMOTE
-
-            data = {
-                "state": (entity.state, str(state)),
-                "attributes": (entity.attributes, attributes),
-                "device_name": (entity.device_name, device),
-            }
-
-            if created or self.entity_manager.compare_data(entity, data):
-                entity.state = state
-                entity.attributes = attributes
-                entity.device_name = device
-
-                entity.set_created_or_updated(created)
-
-            self.entity_manager.set(entity)
-
-        except Exception as ex:
-            self.log_exception(
-                ex, f"Failed to load {DOMAIN_REMOTE}: {entity_name}"
-            )
-
     def _load_binary_sensor_status(
             self,
             device: str,
@@ -266,14 +290,24 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Status"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            system_state = data.get("systemState", {})
+            pws_state = system_state.cleaningMode("pwsState", "off")
+            robot_state = system_state.cleaningMode("robotState", "notConnected")
+            robot_type = system_state.cleaningMode("robotType")
+            is_busy = system_state.cleaningMode("isBusy", False)
+            turn_on_count = system_state.cleaningMode("rTurnOnCount", 0)
+            time_zone = system_state.cleaningMode("timeZone", 0)
+            time_zone_name = system_state.cleaningMode("timeZoneName", "UTC")
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            state = pws_state != STATE_OFF
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Robot": robot_state,
+                "Type": robot_type,
+                "Is Busy": is_busy,
+                "Turn on count": turn_on_count,
+                "Time Zone": f"{time_zone_name} ({time_zone})"
+            }
 
             entity = self.entity_manager.get(DOMAIN_BINARY_SENSOR, entity_name)
             created = entity is None
@@ -313,12 +347,27 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
             day: str,
             data: dict
     ):
+        is_enabled = data.get("isEnabled", False)
+        cleaning_mode = data.get("cleaningMode", {})
+        job_time = data.get("time", {})
+
+        mode = cleaning_mode.get("mode", "all")
+        mode_name = get_cleaning_mode_name(mode)
+        hours = job_time.get("hours", 255)
+        minutes = job_time.get("minutes", 255)
+
         entity_name = f"{device} Schedule {day}"
 
+        job_start_time = None
+        if hours < 255 and minutes < 255:
+            job_start_time = str(datetime.timedelta(hours=hours, minutes=minutes))
+
         try:
-            state = STATE_OFF
+            state = STATE_ON if is_enabled else STATE_OFF
             attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Mode": mode_name,
+                "Start time": job_start_time
             }
 
             # for attr in BINARY_SENSOR_ATTRIBUTES:
@@ -365,14 +414,13 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Filter Status"
 
         try:
-            state = STATE_OFF
+            filter_bag_indication = data.get("filterBagIndication", {})
+            filter_state = filter_bag_indication.cleaningMode("state", 0)
+
+            state = filter_state != 0
             attributes = {
                 ATTR_FRIENDLY_NAME: entity_name
             }
-
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
 
             entity = self.entity_manager.get(DOMAIN_BINARY_SENSOR, entity_name)
             created = entity is None
@@ -406,24 +454,32 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
                 ex, f"Failed to load {DOMAIN_BINARY_SENSOR}: {entity_name}"
             )
 
-    def _load_sensor_connection_type(
+    def _load_binary_sensor_connection(
             self,
             device: str,
             data: dict
     ):
-        entity_name = f"{device} Connection Type"
+        entity_name = f"{device} Connection"
 
         try:
-            state = STATE_OFF
+            system_state = data.get("systemState", {})
+            robot_state = system_state.cleaningMode("robotState", "notConnected")
+
+            debug = data.get("debug", {})
+            wifi_rssi = debug.get("WIFI_RSSI", 0)
+
+            wifi = data.get("wifi", {})
+            net_name = wifi.get("netName")
+
+            state = robot_state != "notConnected"
             attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
+                ATTR_FRIENDLY_NAME: entity_name,
+                "State": robot_state,
+                "RSSI": wifi_rssi,
+                "WIFI Network": net_name
             }
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
-
-            entity = self.entity_manager.get(DOMAIN_SENSOR, entity_name)
+            entity = self.entity_manager.get(DOMAIN_BINARY_SENSOR, entity_name)
             created = entity is None
 
             if created:
@@ -432,13 +488,12 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
                 entity.id = entity_name
                 entity.name = entity_name
                 entity.icon = DEFAULT_ICON
-                entity.domain = DOMAIN_SENSOR
+                entity.domain = DOMAIN_BINARY_SENSOR
                 entity.sensor_device_class = SensorDeviceClass.POWER
                 entity.sensor_state_class = SensorStateClass.MEASUREMENT
 
             data = {
                 "state": (entity.state, str(state)),
-                "attributes": (entity.attributes, attributes),
                 "device_name": (entity.device_name, device),
             }
 
@@ -453,7 +508,7 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
 
         except Exception as ex:
             self.log_exception(
-                ex, f"Failed to load {DOMAIN_SENSOR}: {entity_name}"
+                ex, f"Failed to load {DOMAIN_BINARY_SENSOR}: {entity_name}"
             )
 
     def _load_sensor_cleaning_time(
@@ -464,14 +519,23 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Cleaning Time"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            cycle_info = data.get("cycleInfo", {})
+            cleaning_mode = cycle_info.cleaningMode("cleaningMode", {})
+            mode = cleaning_mode.cleaningMode("mode", "all")
+            mode_name = get_cleaning_mode_name(mode)
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            cycle_time_minutes = cleaning_mode.cleaningMode("cycleTime", 0)
+            cycle_start_time_ts = cycle_info.cleaningMode("cycleStartTime", 0)
+            cycle_start_time = get_date_time_from_timestamp(cycle_start_time_ts)
+
+            cycle_time = str(datetime.timedelta(minutes=cycle_time_minutes))
+
+            state = cycle_time
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Mode": mode_name,
+                "Start Time": cycle_start_time
+            }
 
             entity = self.entity_manager.get(DOMAIN_SENSOR, entity_name)
             created = entity is None
@@ -511,17 +575,30 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
             device: str,
             data: dict
     ):
-        entity_name = f"{device} Filter Status"
+        entity_name = f"{device} Time Left"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            cycle_info = data.get("cycleInfo", {})
+            cleaning_mode = cycle_info.cleaningMode("cleaningMode", {})
+            mode = cleaning_mode.cleaningMode("mode", "all")
+            mode_name = get_cleaning_mode_name(mode)
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            cycle_time = cleaning_mode.cleaningMode("cycleTime", 0)
+            cycle_start_time_ts = cycle_info.cleaningMode("cycleStartTime", 0)
+            cycle_start_time = get_date_time_from_timestamp(cycle_start_time_ts)
+
+            now = datetime.datetime.now()
+            cycle_time_in_seconds = cycle_time * 60
+            since_started = now - cycle_time_in_seconds
+            seconds_left = 0 if since_started > cycle_time_in_seconds else cycle_time_in_seconds - since_started
+
+            state = str(datetime.timedelta(seconds=seconds_left))
+
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Mode": mode_name,
+                "Start Time": cycle_start_time
+            }
 
             entity = self.entity_manager.get(DOMAIN_SENSOR, entity_name)
             created = entity is None
@@ -564,14 +641,24 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Power"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            system_state = data.get("systemState", {})
+            pws_state = system_state.cleaningMode("pwsState", "off")
+            robot_state = system_state.cleaningMode("robotState", "off")
+            robot_type = system_state.cleaningMode("robotType")
+            is_busy = system_state.cleaningMode("isBusy", False)
+            turn_on_count = system_state.cleaningMode("rTurnOnCount", 0)
+            time_zone = system_state.cleaningMode("timeZone", 0)
+            time_zone_name = system_state.cleaningMode("timeZoneName", "UTC")
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            state = pws_state != STATE_OFF
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Robot": robot_state,
+                "Type": robot_type,
+                "Is Busy": is_busy,
+                "Turn on count": turn_on_count,
+                "Time Zone": f"{time_zone_name} ({time_zone})"
+            }
 
             entity = self.entity_manager.get(DOMAIN_SWITCH, entity_name)
             created = entity is None
@@ -612,14 +699,17 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         entity_name = f"{device} Led"
 
         try:
-            state = STATE_OFF
-            attributes = {
-                ATTR_FRIENDLY_NAME: entity_name
-            }
+            led = data.get("led", {})
+            led_mode = led.cleaningMode("ledMode", 1)
+            led_intensity = led.cleaningMode("ledIntensity", 80)
+            led_enable = led.cleaningMode("ledEnable", False)
 
-            # for attr in BINARY_SENSOR_ATTRIBUTES:
-            #    if attr in event_state:
-            #        attributes[attr] = event_state.get(attr)
+            state = led_enable
+            attributes = {
+                ATTR_FRIENDLY_NAME: entity_name,
+                "Mode": led_mode,
+                "Intensity": led_intensity
+            }
 
             entity = self.entity_manager.get(DOMAIN_SWITCH, entity_name)
             created = entity is None
@@ -656,19 +746,19 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
         await self.api.set_cleaning_mode(cleaning_mode)
 
     async def set_delay(self,
+                        device: str,
                         enabled: bool | None = False,
                         mode: str | None = "all",
-                        hours: int | None = 255,
-                        minutes: int | None = 255):
-        await self.api.set_delay(enabled, mode, hours, minutes)
+                        job_time: str | None = None):
+        await self.api.set_delay(device, enabled, mode, job_time)
 
     async def set_schedule(self,
+                           device: str,
                            day: str,
                            enabled: bool | None = False,
                            mode: str | None = "all",
-                           hours: int | None = 255,
-                           minutes: int | None = 255):
-        await self.api.set_schedule(day, enabled, mode, hours, minutes)
+                           job_time: str | None = None):
+        await self.api.set_schedule(device, day, enabled, mode, job_time)
 
     async def set_led_mode(self, mode: int):
         await self.api.set_led_mode(mode)
@@ -679,8 +769,11 @@ class MyDolphinPlusHomeAssistantManager(HomeAssistantManager):
     async def set_led_enabled(self, is_enabled: bool):
         await self.api.set_led_enabled(is_enabled)
 
-    async def drive(self, direction: str):
-        await self.api.drive(direction)
+    async def drive(self, device: str, direction: str):
+        await self.api.drive(device, direction)
+
+    async def pickup(self, device: str):
+        await self.api.pickup(device)
 
     async def set_power_state(self, is_on: bool):
         await self.api.set_power_state(is_on)