+++ custom_components/mydolphin_plus/configuration/helpers/const.py
@@ -5,7 +5,6 @@ Following constants are mandatory for CORE:
     SUPPORTED_PLATFORMS - list of supported HA components to initialize
 """
 from homeassistant.components.binary_sensor import DOMAIN as DOMAIN_BINARY_SENSOR
-from homeassistant.components.remote import DOMAIN as DOMAIN_REMOTE
 from homeassistant.components.select import DOMAIN as DOMAIN_SELECT
 from homeassistant.components.sensor import DOMAIN as DOMAIN_SENSOR
 from homeassistant.components.switch import DOMAIN as DOMAIN_SWITCH
@@ -15,7 +14,6 @@ DEFAULT_NAME = "MyDolphin Plus"
 DOMAIN = "mydolphin_plus"
 SUPPORTED_PLATFORMS = [
     DOMAIN_BINARY_SENSOR,
-    DOMAIN_REMOTE,
     DOMAIN_SELECT,
     DOMAIN_SENSOR,
     DOMAIN_SWITCH